requirements.txt
aiohttp==3.8.4
aiopg==1.4.0
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
contextlib2==21.6.0
databases==0.7.0
fastapi==0.95.0
flake8==6.0.0
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
frozenlist==1.3.3
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.3
httpcore==0.17.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.2
mccabe==0.7.0
msgpack==1.0.5
multidict==4.7.6
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.5.0
psutil==5.9.5
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.6
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyzmq==25.0.2
requests==2.29.0
roundrobin==0.0.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.3.3
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
